+++ bin/handleError.js
@@ -27,6 +27,14 @@ var handlers = {
 
 	DUPLICATE_IMPORT_OPTIONS: function ( err ) {
 		console.error( chalk.red( 'use --input, or pass input path as argument' ) );
+	},
+
+	ROLLUP_WATCH_NOT_INSTALLED: function ( err ) {
+		console.error( chalk.red( 'rollup --watch depends on the rollup-watch package, which could not be found. You can install it globally (recommended) with ' ) + chalk.cyan( 'npm install -g rollup-watch' ) );
+	},
+
+	WATCHER_MISSING_INPUT_OR_OUTPUT: function ( err ) {
+		console.error( chalk.red( 'must specify --input and --output when using rollup --watch' ) );
 	}
 };