+++ src/utils/transform.js
@@ -1,14 +1,5 @@
-import Promise from 'es6-promise/lib/es6-promise/promise.js';
-
 export default function transform ( source, id, transformers ) {
 	let sourceMapChain = [];
-
-	if ( typeof source === 'string' ) {
-		source = {
-			code: source,
-			ast: null
-		};
-	}
 
 	let originalCode = source.code;
 	let ast = source.ast;